Add thrift and tensorpipe backend tests for test_ddp_under_dist_autograd.
Differential Revision: [D22110065](https://our.internmc.facebook.com/intern/diff/D22110065/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D22110065/)!

[ghstack-poisoned]
