+++ test/run_test.py
@@ -77,6 +77,7 @@ TESTS = [
     'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
     'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
     'distributed/rpc/tensorpipe/test_rpc_spawn',
+    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_dist_optimizer_spawn',
     'distributed/rpc/test_rpc_spawn',
@@ -98,6 +99,7 @@ WINDOWS_BLACKLIST = [
     'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
     'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
     'distributed/rpc/tensorpipe/test_rpc_spawn',
+    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_dist_optimizer_spawn',
     'distributed/rpc/test_rpc_spawn',
@@ -115,6 +117,7 @@ ROCM_BLACKLIST = [
     'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
     'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
     'distributed/rpc/tensorpipe/test_rpc_spawn',
+    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_optimizer_spawn',
@@ -160,6 +163,7 @@ SLOW_TESTS = [
     'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
     'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
     'distributed/rpc/tensorpipe/test_rpc_spawn',
+    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_rpc_spawn',
     'distributed/test_ddp_under_dist_autograd',
